Fix some portability issues with the testsuite.
Some Linux distros don't provide libc.a by default and/or don't use
Bash as default shell.

+++ tests/test-1743dd3d.sh
@@ -1,3 +1,7 @@
+if [ ! -x ${ROOTFS}/bin/true ] || [ -z `which mcookie` ] || [ -z `which mkdir` ] || [ -z `which echo` ] || [ -z `which chmod` ]; then
+    exit 125;
+fi
+
 TMP=/tmp/`mcookie`
 rm -f ${ROOTFS}/${TMP}
 
@@ -9,3 +13,5 @@ chmod -x ${ROOTFS}/${TMP}
 
 chmod +x ${ROOTFS}/${TMP}
 ${PROOT} ${ROOTFS} ${TMP}
+
+rm -f ${ROOTFS}/${TMP}
