+++ tests/test-ebadf0dd.sh
@@ -1,3 +1,7 @@
+if [ -z `which mcookie` ] || [ -z `which rm` ] || [ ! -x  ${ROOTFS}/bin/true ]; then
+    exit 125;
+fi
+
 TMP=/tmp/`mcookie`
 rm -fr ${ROOTFS}/${TMP}
